backend/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
python-multipart==0.0.6
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pydantic==1.10.12
httplib2==0.22.0
Pillow==10.1.0

